requirements.txt
django
python-decouple
pytube
ffmpeg-python